+++ spec/rubocop/cop/style/space_inside_hash_literal_braces_spec.rb
@@ -127,6 +127,68 @@ describe RuboCop::Cop::Style::SpaceInsideHashLiteralBraces, :config do
     end
   end
 
+  context 'when EnforcedStyle is compact' do
+    let(:cop_config) { { 'EnforcedStyle' => 'compact' } }
+
+    it "doesn't register an offense for non-nested hashes with spaces" do
+      inspect_source(cop, 'h = { a: 1, b: 2 }')
+      expect(cop.offenses).to be_empty
+      expect(cop.config_to_allow_offenses).to eq('EnforcedStyle' => 'compact')
+    end
+
+    it 'registers an offense for nested hashes with spaces' do
+      inspect_source(cop, 'h = { a: { a: 1, b: 2 } }')
+      expect(cop.offenses.size).to eq 1
+      expect(cop.messages).to eq(['Space inside } detected.'])
+    end
+
+    it 'registers an offense for opposite + correct' do
+      inspect_source(cop,
+                     'h = {a: 1 }')
+      expect(cop.messages).to eq(['Space inside { missing.'])
+      expect(cop.config_to_allow_offenses).to eq('Enabled' => false)
+    end
+
+    it 'auto-corrects hashes with no space' do
+      new_source = autocorrect_source(cop, ['h = {a: 1, b: 2}',
+                                            'h = {a => 1 }'])
+      expect(new_source).to eq(['h = { a: 1, b: 2 }',
+                                'h = { a => 1 }'].join("\n"))
+    end
+
+    it 'auto-corrects nested hashes with spaces' do
+      new_source = autocorrect_source(cop, ['h = { a: { a: 1, b: 2 } }',
+                                            'h = {a => method { 1 } }'])
+      expect(new_source).to eq(['h = { a: { a: 1, b: 2 }}',
+                                'h = { a => method { 1 }}'].join("\n"))
+    end
+
+    it 'registers offenses for hashes with no spaces' do
+      inspect_source(cop,
+                     ['h = {a: 1, b: 2}',
+                      'h = {a => 1}'])
+      expect(cop.offenses.size).to eq 4
+    end
+
+    it 'accepts multiline hash' do
+      inspect_source(cop,
+                     ['h = {',
+                      '      a: 1,',
+                      '      b: 2,',
+                      '}'])
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'accepts multiline hash with comment' do
+      inspect_source(cop,
+                     ['h = { # Comment',
+                      '      a: 1,',
+                      '      b: 2,',
+                      '}'])
+      expect(cop.offenses).to be_empty
+    end
+  end
+
   it 'accepts hashes with spaces by default' do
     inspect_source(cop,
                    ['h = { a: 1, b: 2 }',